+++ src/com/cx/myobject/MyNoteDialog.java
@@ -33,8 +33,8 @@ private Button dialogSaveBtn,dialogCancelBtn;
 		
 		dialogSaveBtn=(Button)view.findViewById(R.id.dialog_save_btn);
 		dialogCancelBtn=(Button)view.findViewById(R.id.dialog_cancel_btn);
-		int dialogWidth=(int)(screenWidth*3/4);
-		int dialogHeight=(int)(dialogWidth*2/3);
+		int dialogWidth=screenWidth*3/4;
+		int dialogHeight=dialogWidth*2/3;
 		dialogEt.getLayoutParams().height=dialogHeight*1/3;
 		super.addContentView(view, new LayoutParams(dialogWidth,dialogHeight));
 	}
